requirements.txt
fastapi~=0.115.14
uvicorn~=0.35.0
python-multipart
easyocr~=1.7.2
pymupdf
google-genai
python-dotenv~=1.1.1
requests~=2.32.4
streamlit~=1.46.1
transformers~=4.53.0
torch
pydantic~=2.11.7
pandas~=2.3.0
python-docx~=1.2.0
protobuf~=6.31.1
faiss-cpu~=1.11.0
sentence-transformers~=5.0.0
h11~=0.16.0
pip~=25.1.1
attrs~=25.3.0
filelock~=3.18.0
sniffio~=1.3.1
typing_extensions~=4.14.0
starlette~=0.46.2
pydantic_core~=2.33.2
websockets~=15.0.1
anyio~=4.9.0
click~=8.2.1
annotated-types~=0.7.0
PyYAML~=6.0.2
Jinja2~=3.1.6
httpx~=0.28.1
typing_extensions~=4.14.0
pydantic_core~=2.33.2